allRepos.txt: update mr setup instructions
It seems that when chaning from a global mr config file, when another config is chained, mr forgets to update the $MR_CONFIG variable, which helps to resolve the current path of the running config.
Remove instructions for a global mr config+chaining, just use it locally.

Thanks to Bernd Dorer for reporting the problem.